TeensyRWR: tweaks

+++ src/SimLinkup/HardwareSupport/TeensyRWR/DrawPoint.cs
@@ -10,8 +10,8 @@ namespace SimLinkup.HardwareSupport.TeensyRWR
 
         public static implicit operator uint(DrawPoint drawPoint)
         {
-            var xBits = ((uint)drawPoint.X & 0xFFF) << 12;
-            var yBits = ((uint)drawPoint.Y & 0xFFF);
+            var xBits = ((uint)(((uint)drawPoint.X) & 0xFFF) << 12);
+            var yBits = (((uint)drawPoint.Y) & 0xFFF);
             var zBit = (uint) (drawPoint.BeamOn ? 1 : 0) << 24;
             return xBits | yBits | zBit;
         }
